+++ packages/form/src/BaseForm/index.tsx
@@ -18,7 +18,11 @@ import Submitter, { SubmitterProps } from '../components/Submitter';
 import { GroupProps, FieldProps } from '../interface';
 
 export interface CommonFormProps {
-  submitter?: Omit<SubmitterProps, 'form'> | false;
+  submitter?:
+    | SubmitterProps<{
+        form?: FormInstance<any>;
+      }>
+    | false;
 
   /**
    * @name 表单结束后调用
@@ -35,7 +39,7 @@ export interface CommonFormProps {
 export interface BaseFormProps extends FormProps, CommonFormProps {
   contentRender?: (
     items: React.ReactNode[],
-    submitter: ReactElement<Omit<SubmitterProps, 'form'>> | undefined,
+    submitter: ReactElement<SubmitterProps> | undefined,
     form: FormInstance<any>,
   ) => React.ReactNode;
   fieldProps?: FieldProps;
@@ -84,7 +88,7 @@ const BaseForm: React.FC<BaseFormProps> = (props) => {
   const [, updateState] = useState(false);
 
   const items = React.Children.toArray(children);
-  const submitterProps: Omit<SubmitterProps, 'form'> =
+  const submitterProps: SubmitterProps =
     typeof submitter === 'boolean' || !submitter ? {} : submitter;
 
   /**