+++ packages/table/src/component/useEditable.tsx
@@ -32,12 +32,31 @@ export interface TableRowEditable<T> {
   /**
    * 行保存的时候
    */
-  onSave?: (key: React.Key, row: T & { index: number }) => Promise<void>;
+  onSave?: (
+    key: React.Key,
+    row: T & { index: number },
+    newLineConfig?: {
+      row: T | undefined;
+      options: AddLineOptions;
+    },
+  ) => Promise<boolean | void>;
+
+  /**
+   * 行保存的时候
+   */
+  onCancel?: (
+    key: React.Key,
+    row: T & { index: number },
+    newLineConfig?: {
+      row: T | undefined;
+      options: AddLineOptions;
+    },
+  ) => Promise<boolean | void>;
 
   /**
    * 行删除的时候
    */
-  onDelete?: (key: React.Key, row: T & { index: number }) => Promise<void>;
+  onDelete?: (key: React.Key, row: T & { index: number }) => Promise<boolean | void>;
 
   /**
    * 删除行时的确认消息
@@ -60,10 +79,14 @@ export type ActionRenderConfig<T> = {
   form: FormInstance<any>;
   cancelEditable: (key: React.Key) => void;
   onSave: TableRowEditable<T>['onSave'];
+  onCancel: TableRowEditable<T>['onCancel'];
   onDelete: TableRowEditable<T>['onDelete'];
   deletePopconfirmMessage: TableRowEditable<T>['deletePopconfirmMessage'];
   setEditableRowKeys: (value: React.Key[]) => void;
-  isNewLine: boolean;
+  newLineConfig?: {
+    row: T | undefined;
+    options: AddLineOptions;
+  };
 };
 
 /**
@@ -160,12 +183,16 @@ function editableRowByKey<RecordType>(
   return source;
 }
 
+/**
+ * 保存按钮的dom
+ * @param ActionRenderConfig
+ */
 const SaveEditableAction: React.FC<ActionRenderConfig<any> & { row: any }> = ({
   recordKey,
   onSave,
   form,
   row,
-  cancelEditable,
+  newLineConfig,
 }) => {
   const [loading, setLoading] = useState<boolean>(false);
   return (
@@ -179,13 +206,15 @@ const SaveEditableAction: React.FC<ActionRenderConfig<any> & { row: any }> = ({
             recursive: true,
           });
           const fields = form.getFieldValue([recordKey]);
-          await onSave?.(recordKey, { ...row, ...fields });
+          const success = await onSave?.(recordKey, { ...row, ...fields }, newLineConfig);
+          if (success === false) {
+            setLoading(false);
+            return;
+          }
           form.resetFields([recordKey]);
-          setLoading(false);
-          setTimeout(() => {
-            cancelEditable(recordKey);
-          }, 0);
         } catch (e) {
+          // eslint-disable-next-line no-console
+          console.log(e);
           setLoading(false);
         }
       }}
@@ -202,6 +231,10 @@ const SaveEditableAction: React.FC<ActionRenderConfig<any> & { row: any }> = ({
   );
 };
 
+/**
+ * 删除按钮 dom
+ * @param ActionRenderConfig
+ */
 const DeleteEditableAction: React.FC<ActionRenderConfig<any> & { row: any }> = ({
   recordKey,
   onDelete,
@@ -213,8 +246,9 @@ const DeleteEditableAction: React.FC<ActionRenderConfig<any> & { row: any }> = (
   const onConfirm = async () => {
     try {
       setLoading(true);
-      await onDelete?.(recordKey, row);
+      const success = await onDelete?.(recordKey, row);
       setLoading(false);
+      if (success === false) return;
       setTimeout(() => {
         cancelEditable(recordKey);
       }, 0);
@@ -239,13 +273,16 @@ const DeleteEditableAction: React.FC<ActionRenderConfig<any> & { row: any }> = (
 };
 
 const defaultActionRender: ActionRenderFunction<any> = (row, config) => {
-  const { recordKey, isNewLine, cancelEditable } = config;
+  const { recordKey, newLineConfig, form, onCancel, cancelEditable } = config;
   return [
     <SaveEditableAction key="save" {...config} row={row} />,
-    !isNewLine && <DeleteEditableAction key="delete" {...config} row={row} />,
+    !newLineConfig && <DeleteEditableAction key="delete" {...config} row={row} />,
     <a
       key="cancel"
-      onClick={() => {
+      onClick={async () => {
+        const fields = form.getFieldValue([recordKey]);
+        const success = await onCancel?.(recordKey, fields, newLineConfig);
+        if (success === false) return;
         cancelEditable(recordKey);
       }}
     >
@@ -285,6 +322,8 @@ function useEditable<RecordType>(
       }
     | undefined
   >(undefined);
+
+  // 这里这么做是为了存上次的状态，不然每次存一下再拿
   newLineRecordRef.current = newLineRecord;
 
   const editableType = props.type || 'single';
@@ -372,7 +411,23 @@ function useEditable<RecordType>(
         recordKey: key,
         cancelEditable,
         index: row.index,
-        isNewLine: !!newLineRecord,
+        newLineConfig: newLineRecord,
+        onCancel: async (
+          recordKey: React.Key,
+          editRow: RecordType & {
+            index: number;
+          },
+          isNewLine?: {
+            row: RecordType | undefined;
+            options: AddLineOptions;
+          },
+        ) => {
+          const success = await props?.onCancel?.(recordKey, editRow, isNewLine);
+          if (success === false) {
+            return false;
+          }
+          return true;
+        },
         onDelete: async (
           recordKey: React.Key,
           editRow: RecordType & {
@@ -386,24 +441,36 @@ function useEditable<RecordType>(
             key: recordKey,
             childrenColumnName: props.childrenColumnName || 'children',
           };
-          await props?.onDelete?.(recordKey, editRow);
+          const success = await props?.onDelete?.(recordKey, editRow);
+          if (success === false) {
+            return false;
+          }
           props.setDataSource(editableRowByKey(actionProps, 'delete'));
+          return true;
         },
         onSave: async (
           recordKey: React.Key,
           editRow: RecordType & {
             index: number;
           },
+          isNewLine?: {
+            row: RecordType | undefined;
+            options: AddLineOptions;
+          },
         ) => {
-          const { options } = newLineRecordRef.current || {};
-          await props?.onSave?.(recordKey, editRow);
-          if (newLineRecordRef.current && recordKey === options?.recordKey) {
+          const { options } = isNewLine || {};
+          const success = await props?.onSave?.(recordKey, editRow, isNewLine);
+          if (success === false) {
+            return false;
+          }
+          cancelEditable(recordKey);
+          if (isNewLine) {
             if (options?.position === 'top') {
               props.setDataSource([editRow, ...props.dataSource]);
             } else {
               props.setDataSource([...props.dataSource, editRow]);
             }
-            return;
+            return true;
           }
           const actionProps = {
             data: props.dataSource,
@@ -413,6 +480,7 @@ function useEditable<RecordType>(
             childrenColumnName: props.childrenColumnName || 'children',
           };
           props.setDataSource(editableRowByKey(actionProps, 'update'));
+          return true;
         },
         form,
         editableKeys,