+++ packages/table/src/typing.ts
@@ -10,9 +10,10 @@ import {
   ProSchema,
   ProSchemaComponentTypes,
   SearchTransformKeyFn,
+  ProTableEditableFnType,
 } from '@ant-design/pro-utils';
 import { CardProps } from 'antd/lib/card';
-import { FormItemProps } from 'antd/lib/form';
+import { FormInstance, FormItemProps } from 'antd/lib/form';
 import { SpinProps } from 'antd/lib/spin';
 import { TableProps } from 'antd/lib/table';
 
@@ -130,10 +131,18 @@ export type ProColumnType<T = unknown> = ProSchema<
      * form 的排序
      */
     order?: number;
+
     /**
      * 传给 Form.Item 的 props
      */
-    formItemProps?: Partial<Omit<FormItemProps, 'children'>>;
+    formItemProps?:
+      | ((form: FormInstance<any>) => Partial<Omit<FormItemProps, 'children'>>)
+      | Partial<Omit<FormItemProps, 'children'>>;
+
+    /**
+     * 可编辑表格是否可编辑
+     */
+    editable?: boolean | ProTableEditableFnType<T>;
   },
   Partial<ActionType>
 >;