+++ packages/form/CHANGELOG.md
@@ -2,6 +2,34 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.8.0](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.6...@ant-design/pro-form@1.8.0) (2020-12-08)
+
+### Features
+
+- **table:** fieldProps support function ([#1227](https://github.com/ant-design/pro-components/issues/1227)) ([bc34fb0](https://github.com/ant-design/pro-components/commit/bc34fb0ce11d006c83b06166f6cf8903bd5d3e1a))
+
+## [1.7.6](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.5...@ant-design/pro-form@1.7.6) (2020-12-07)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [1.7.5](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.4...@ant-design/pro-form@1.7.5) (2020-12-04)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [1.7.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.3...@ant-design/pro-form@1.7.4) (2020-12-03)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [1.7.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.2...@ant-design/pro-form@1.7.3) (2020-12-03)
+
+**Note:** Version bump only for package @ant-design/pro-form
+
+## [1.7.2](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.1...@ant-design/pro-form@1.7.2) (2020-12-02)
+
+### Bug Fixes
+
+- **form:** fix optionRender no has form error ([#1147](https://github.com/ant-design/pro-components/issues/1147)) ([3398fb9](https://github.com/ant-design/pro-components/commit/3398fb999952eebf6b9ba5736e6dc5d44a5fecce))
+
 ## [1.7.1](https://github.com/ant-design/pro-components/compare/@ant-design/pro-form@1.7.0...@ant-design/pro-form@1.7.1) (2020-11-30)
 
 ### Bug Fixes