+++ packages/layout/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.5.19](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.5.18...@ant-design/pro-layout@6.5.19) (2020-12-08)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
+## [6.5.18](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.5.17...@ant-design/pro-layout@6.5.18) (2020-12-07)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
+## [6.5.17](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.5.16...@ant-design/pro-layout@6.5.17) (2020-12-02)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
 ## [6.5.16](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.5.15...@ant-design/pro-layout@6.5.16) (2020-11-27)
 
 **Note:** Version bump only for package @ant-design/pro-layout